==> 04-lisbon-cicleways.ipynb <==
#### Import packages

import geojson
import json
import folium
import geopandas as gpd
import pandas as pd

#### Open Lisbon's municipality data on cycleways

with open('Data/Ciclovias.geojson') as f:
    ciclovias = geojson.load(f)

#### Correct some mistakes in the data

for i in range(len(ciclovias['features'])):
    ciclovia = ciclovias['features'][i]
    ciclovia['IDunico'] = ciclovia['properties']['COD_SIG']
      
    if ciclovia['IDunico'] == '0000455':
        ciclovia['properties']['ANO'] = 2000

    if ciclovia['IDunico'] == '0000371':
        ciclovia['properties']['ANO'] = 2001
        
    if ciclovia['IDunico'] == '6708':
        ciclovia['properties']['ANO'] = 2003
        
    if ciclovia['IDunico'] == '0000204' or \
        ciclovia['IDunico'] == '0000080' or \
        ciclovia['IDunico'] == '0000205':
        ciclovia['properties']['ANO'] = 2005
        
    if ciclovia['IDunico'] == '0000025' or \
        ciclovia['IDunico'] == '0000173' or \
        ciclovia['IDunico'] == '0000085':
        ciclovia['properties']['ANO'] = 2008
    if ciclovia['IDunico'] == '0000224' or \
        ciclovia['IDunico'] == '000002' or \
        ciclovia['IDunico'] == '0000015' or \
        ciclovia['IDunico'] == '0000012' or \
        ciclovia['IDunico'] == '0000138' or \
        ciclovia['IDunico'] == '0000144' or \
        ciclovia['IDunico'] == '0000237' or \
        ciclovia['IDunico'] == '0000145' or \
        ciclovia['IDunico'] == '0000225' or \
        ciclovia['IDunico'] == '0000132' or \
        ciclovia['IDunico'] == '0000205' or \
        ciclovia['IDunico'] == '0000174' or \
        ciclovia['IDunico'] == '0000024' or \
        ciclovia['IDunico'] == '0000210' or \
        ciclovia['IDunico'] == '0000081' or \
        ciclovia['IDunico'] == '0000028' or \
        ciclovia['IDunico'] == '0000010' or \
        ciclovia['IDunico'] == '0000011' or \
        ciclovia['IDunico'] == '0000214' or \
        ciclovia['IDunico'] == '0000013' or \
        ciclovia['IDunico'] == '0000212' or \
        ciclovia['IDunico'] == '000008' or \
        ciclovia['IDunico'] == '0000216' or \
        ciclovia['IDunico'] == '0000009' or \
        ciclovia['IDunico'] == '0000202' or \
        ciclovia['IDunico'] == '0000184' or \
        ciclovia['IDunico'] == '000007':
        ciclovia['properties']['ANO'] = 2009
    if ciclovia['IDunico'] == '0000036' or \
        ciclovia['IDunico'] == '0000026' or \
        ciclovia['IDunico'] == '14296' or \
        ciclovia['IDunico'] == '000019' or \
        ciclovia['IDunico'] == '152515' or \
        ciclovia['IDunico'] == '0000006' or \
        ciclovia['IDunico'] == '12211' or \
        ciclovia['IDunico'] == '3185' or \
        ciclovia['IDunico'] == '152515' or \
        ciclovia['IDunico'] == '23242' or \
        ciclovia['IDunico'] == '0000285' or \
        ciclovia['IDunico'] == '159396' or \
        ciclovia['IDunico'] == '0000284' or \
        ciclovia['IDunico'] == '159396' or \
        ciclovia['IDunico'] == '159399' or \
        ciclovia['IDunico'] == '0000041' or \
        ciclovia['IDunico'] == '19676' or \
        ciclovia['IDunico'] == '12210' or \
        ciclovia['IDunico'] == '14286' or \
        ciclovia['IDunico'] == '000016' or \
        ciclovia['IDunico'] == '0000061' or \
        ciclovia['IDunico'] == '0000064' or \
        ciclovia['IDunico'] == '0000065' or \
        ciclovia['IDunico'] == '0000066' or \
        ciclovia['IDunico'] == '0000063' or \
        ciclovia['IDunico'] == '0000062' or \
        ciclovia['IDunico'] == '6244' or \
        ciclovia['IDunico'] == '6244' or \
        ciclovia['IDunico'] == '000017':
        ciclovia['properties']['ANO'] = 2010
    if ciclovia['IDunico'] == '0000017' or \
        ciclovia['IDunico'] == '0000018' or \
        ciclovia['IDunico'] == '0000146' or \
        ciclovia['IDunico'] == '0000181' or \
        ciclovia['IDunico'] == '0000147':
        ciclovia['properties']['ANO'] = 2011
    if ciclovia['IDunico'] == '0000266' or \
        ciclovia['IDunico'] == '0000265' or \
        ciclovia['IDunico'] == '0000028' or \
        ciclovia['IDunico'] == '0000188' or \
        ciclovia['IDunico'] == '0000078' or \
        ciclovia['IDunico'] == '0000037':
        ciclovia['properties']['ANO'] = 2012
        
    if ciclovia['IDunico'] == '18320' or \
        ciclovia['IDunico'] == '159178' or \
        ciclovia['IDunico'] == '159044' or \
        ciclovia['IDunico'] == '158526' or \
        ciclovia['IDunico'] == '159033' or \
        ciclovia['IDunico'] == '159043' or \
        ciclovia['IDunico'] == '159023' or \
        ciclovia['IDunico'] == '159031' or \
        ciclovia['IDunico'] == '159021' or \
        ciclovia['IDunico'] == '159179' or \
        ciclovia['IDunico'] == '159022':
        ciclovia['properties']['ANO'] = 2013

    if ciclovia['IDunico'] == '0000045' or \
        ciclovia['IDunico'] == '0000140' or \
        ciclovia['IDunico'] == '15102' or \
        ciclovia['IDunico'] == '2607' or \
        ciclovia['IDunico'] == '14168' or \
        ciclovia['IDunico'] == '0000067' or \
        ciclovia['IDunico'] == '0000045' or \
        ciclovia['IDunico'] == '0000367' or \
        ciclovia['IDunico'] == '0000045' or \
        ciclovia['IDunico'] == 'B0000367':
        ciclovia['properties']['ANO'] = 2014

    if ciclovia['IDunico'] == '5777' or \
        ciclovia['IDunico'] == '5684' or \
        ciclovia['IDunico'] == '0000215' or \
        ciclovia['IDunico'] == '5593' or \
        ciclovia['IDunico'] == '0000243' or \
        ciclovia['IDunico'] == '0000057' or \
        ciclovia['IDunico'] == 'B0000038' or \
        ciclovia['IDunico'] == 'B0000183' or \
        ciclovia['IDunico'] == '0000240':
        ciclovia['properties']['ANO'] = 2015

    if ciclovia['IDunico'] == '18155' or \
        ciclovia['IDunico'] == '17340' or \
        ciclovia['IDunico'] == '158357' or \
        ciclovia['IDunico'] == '18153' or \
        ciclovia['IDunico'] == '17343' or \
        ciclovia['IDunico'] == '18029' or \
        ciclovia['IDunico'] == '0000077' or \
        ciclovia['IDunico'] == '18031' :
        ciclovia['properties']['ANO'] = 2016

    if ciclovia['IDunico'] == '0000373':
        ciclovia['properties']['ANO'] = 2018
        
    if ciclovia['IDunico'] == '0000372' or \
        ciclovia['IDunico'] == '0000392' or \
        ciclovia['IDunico'] == '0000391' or \
        ciclovia['IDunico'] == '4198' or \
        ciclovia['IDunico'] == '15403' or \
        ciclovia['IDunico'] == '153668' or \
        ciclovia['IDunico'] == '4128' or \
        ciclovia['IDunico'] == '4671' or \
        ciclovia['IDunico'] == '156600' or \
        ciclovia['IDunico'] == '4614' or \
        ciclovia['IDunico'] == '4198' or \
        ciclovia['IDunico'] == '0000325' or \
        ciclovia['IDunico'] == '0000325' or \
        ciclovia['IDunico'] == '15403' or \
        ciclovia['IDunico'] == '15620' or \
        ciclovia['IDunico'] == '0000393' or \
        ciclovia['IDunico'] == '0000390':
        ciclovia['properties']['ANO'] = 2019   

### Plot all cycleways as of 2020

gdf_freguesias = gpd.read_file("Data/Lisboa_Freguesias/Lisboa_Freguesias_CAOP2015_TM06.shp")
gdf_freguesias = gdf_freguesias.to_crs(epsg=4326)
gdf_freguesias.geometry.index = gdf_freguesias['DICOFRE']

style_1 = {'fillColor': '#1ED674', 'color': '#1ED674'}
style_2 = {'fillColor': '#D6A86E', 'color': '#D6A86E'}
style_3 = {'fillColor': '#233BF7', 'color': '#233BF7'}

unique_zones = gdf_freguesias['DICOFRE'].unique()
m = folium.Map([38.748662, -9.145801],
                 zoom_start=12,
                 tiles='cartodbpositron')

for i in range(len(ciclovias['features'])):
    ciclovia = ciclovias['features'][i]
    
    try:
        ano = int(ciclovia['properties']['ANO'])
        if ano <= 2020:

            folium.GeoJson(
                ciclovia,
                name='geojson',
                style_function=lambda x:style_1
            ).add_to(m)
    except ValueError:
        folium.GeoJson(
                ciclovia,
                name='geojson',
                style_function=lambda x:style_3
            ).add_to(m)
    
for i, zone in enumerate(unique_zones):
    mask = gdf_freguesias['DICOFRE'] == zone

    example_or = gdf_freguesias.loc[mask]
    folium.GeoJson(example_or,name='polygon_or',style_function=lambda x:style_2).add_to(m)

m

### Create maps of Lisbon's cycleways over time

gdf_freguesias = gpd.read_file("Data/Lisboa_Freguesias/Lisboa_Freguesias_CAOP2015_TM06.shp")
gdf_freguesias = gdf_freguesias.to_crs(epsg=4326)
gdf_freguesias.geometry.index = gdf_freguesias['DICOFRE']
style_0 = {'fillColor': '#D6A86E', 'color': '#D6A86E'}
style_1 = {'fillColor': '#1ED674', 'color': '#1ED674'}
style_2 = {'fillColor': '#5CC1F7', 'color': '#5CC1F7'}
style_3 = {'fillColor': '#472EAB', 'color': '#472EAB'}
style_4 = {'fillColor': '#AB2E6E', 'color': '#AB2E6E'}

unique_zones = gdf_freguesias['DICOFRE'].unique()


for current_year in range(2013, 2021):

    m = folium.Map([38.747156, -9.156788],
                 zoom_start=13,
                 tiles='cartodbpositron')
    
    for i in range(len(ciclovias['features'])):
        ciclovia = ciclovias['features'][i]

        try:
            ano = int(ciclovia['properties']['ANO'])
            if ano <= current_year:
                folium.GeoJson(
                    ciclovia,
                    name='geojson',
                    style_function=lambda x:style_3
                ).add_to(m)

        except ValueError:
            pass
    
    m.save(str(current_year)+".html")



### CASE STUDY: Marvila

gdf_freguesias = gpd.read_file("Data/Lisboa_Freguesias/Lisboa_Freguesias_CAOP2015_TM06.shp")
gdf_freguesias = gdf_freguesias.to_crs(epsg=4326)
gdf_freguesias.geometry.index = gdf_freguesias['DICOFRE']
style_0 = {'fillColor': '#D6A86E', 'color': '#D6A86E'}
style_1 = {'fillColor': '#1ED674', 'color': '#1ED674'}
style_2 = {'fillColor': '#5CC1F7', 'color': '#5CC1F7'}
style_3 = {'fillColor': '#472EAB', 'color': '#472EAB'}
style_4 = {'fillColor': '#AB2E6E', 'color': '#AB2E6E'}

unique_zones = gdf_freguesias['DICOFRE'].unique()
m = folium.Map([38.749906, -9.110679],
                 zoom_start=14,
                 tiles='cartodbpositron')

for i in range(len(ciclovias['features'])):
    ciclovia = ciclovias['features'][i]
    
    try:
        ano = int(ciclovia['properties']['ANO'])
        if ano <= 2016:
            folium.GeoJson(
                ciclovia,
                name='geojson',
                style_function=lambda x:style_3
            ).add_to(m)
        if ano == 2017:
            folium.GeoJson(
                ciclovia,
                name='geojson',
                style_function=lambda x:style_1
            ).add_to(m)
        if ano == 2018:
            folium.GeoJson(
                ciclovia,
                name='geojson',
                style_function=lambda x:style_2
            ).add_to(m)
    except ValueError:
        pass
    
for i, zone in enumerate(unique_zones):
    if zone != '110621':
        continue
    mask = gdf_freguesias['DICOFRE'] == zone

    example_or = gdf_freguesias.loc[mask]
    folium.GeoJson(example_or,name='polygon_or',style_function=lambda x:style_0).add_to(m)

m
