==> 05-osm_data_road_length.ipynb <==
import osmium as osm
import pandas as pd

## Total road length in Lisbon over time

class RoadLengthHandler(osm.SimpleHandler):
    def __init__(self):
        super(RoadLengthHandler, self).__init__()
        self.length_road = 0.0
        self.length = 0.0
        self.length_cycle = 0.0
        
    def way(self, w):
        # print(w)
        if 'highway' in w.tags:
            try:
                self.length += osm.geom.haversine_distance(w.nodes)
            except osm.InvalidLocationError:
                # A location error might occur if the osm file is an extract
                # where nodes of ways near the boundary are missing.
                #print("WARNING: way %d incomplete. Ignoring." % w.id)
                pass
            
            
            if w.tags['highway'] == 'motorway':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'trunk':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'primary':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'secondary':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'terciary':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'motorway_link':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'trunk_link':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'primary_link':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'secondary_link':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'tertiary_link':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'road':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'residential':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'unclassified':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'living_street':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
            if w.tags['highway'] == 'busway':
                try:
                    self.length_road += osm.geom.haversine_distance(w.nodes)
                except osm.InvalidLocationError:
                    pass
        

for i in range(2013, 2021):
    h = RoadLengthHandler()
    year = str(i)
    
    h.apply_file("Data/lisbon_"+ year + ".osm.pbf", locations=True, idx='flex_mem')
    # print(year + ' Total way length: %.2f km' % (h.length/1000))
    print(year + ' Total way length_road: %.2f km' % (h.length_road/1000))
